+++ app.rb
@@ -4,6 +4,9 @@ require_relative 'student'
 require_relative 'teacher'
 require_relative 'list_books'
 require_relative 'list_people'
+require_relative 'create_book'
+require_relative 'create_rental'
+require_relative 'list_rentals'
 
 class App
   attr_reader :people
@@ -53,46 +56,5 @@ class App
     answer = gets.chomp.to_i
     exit if answer == 7
     option_checker(answer)
-  end
-
-  def create_book
-    print 'Title: '
-    title = gets.chomp
-    print 'Author: '
-    author = gets.chomp
-    book = Book.new(title, author)
-    @books.push(book)
-    puts 'Book is created successfully'
-    run
-  end
-
-  def create_rental
-    puts 'Select the book from the following list by number (not ID)'
-    @books.each_with_index do |book, idx|
-      puts "#{idx}) Title: '#{book.title}', Author: '#{book.author}'"
-    end
-    selected_id_book = gets.chomp.to_i
-    print 'Date: '
-    date = gets.chomp
-
-    puts 'Select a person from the following list by number (not ID)'
-    @people.each_with_index do |person, idx|
-      puts "#{idx}) [#{person.class.name}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
-    end
-    selected_id_person = gets.chomp.to_i
-
-    rental = Rental.new(date, @books[selected_id_book], @people[selected_id_person])
-    @rentals.push(rental)
-    puts 'Rental is created successfully'
-    run
-  end
-
-  def list_rentals
-    print 'ID of person: '
-    id = gets.chomp.to_i
-    rentals = @rentals.select { |rental| id == rental.person.id }
-
-    rentals.each { |item| puts "Date: #{item.date},  Book #{item.book.title}, by #{item.book.author}" }
-    run
   end
 end
